+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - added AIDER_DESK_AIDER_VERSION environment variable to use custom version of Aider
 - fixed claude-4-sonnet-20250514 default model
 - added caching info for Requesty provider
+- setting to use Vim bindings in Prompt field
 
 ## [0.17.0]